kms/health: move options to its own file
